Create wjohns53.md
Hello everybody. My name is William Johnson and I am an undergraduate student pursuing in a computer science major and a cybersecurity minor. I picked this course out of curiosity as archaeology sounded cool, and this course is for my cybersecurity minor. For hobby, I enjoy web development which includes NodeJS, ReactJS, Redux, GraphQL, and Redis.